♻️ Refactor validation.js
데이터와 액션과 연산을 용도에 맞게 모듈화하였습니다.
