+++ login/index.html
@@ -50,6 +50,6 @@
         </section>
       </main>
     </div>
-    <script src="/scripts/validation.js"></script>
+    <script type="module" src="/scripts/validation.js"></script>
   </body>
 </html>
