+++ scripts/constants/validityMessage.js
@@ -1,0 +1,12 @@
+export const validityMessage = {
+  emailIsEmpty: `이메일을 입력해주세요.`,
+  invalidEmail: `잘못된 이메일 형식입니다.`,
+  emailIsExist: `사용 중인 이메일입니다.`,
+  nicknameIsEmpty: `닉네임을 입력해주세요.`,
+  invalidNickname: `닉네임은 한글, 영문, 숫자로 2~16자 사이로 입력하세요`,
+  passwordIsEmpty: `비밀번호를 입력해주세요.`,
+  passwordIsTooShort: `비밀번호를 8자리 이상 입력해주세요.`,
+  passwordIsNotMatch: `비밀번호가 일치하지 않습니다.`,
+  wrongPassword: `이메일 또는 비밀번호가 일치하지 않습니다.`,
+  signupSuccess: `회원가입이 완료되었습니다. 로그인 페이지로 이동합니다.`,
+};
