+++ signup/index.html
@@ -63,6 +63,6 @@
         </section>
       </main>
     </div>
-    <script src="/scripts/validation.js"></script>
+    <script type="module" src="/scripts/validation.js"></script>
   </body>
 </html>
